move character generation to same view folder so it's not floating
